Detection faces on camera(real time)
VisageVision is a recent project that showcases a Face Detection Camera Application. This project has garnered interest due to its real-world applicability and advanced computer vision techniques. When a human face enters the frame of the camera, VisageVision detects it with a square shape using the OpenCV library, an open-source computer vision and machine learning software library that provides a wide range of tools.